Try using mysql_config if its available.
As there are different versions of mysql_config depending on the version
we test to see what works. This seems to be mostly ok for the current
platforms we build on but we hope for a somewhat stable ABI of the
mysql_config program in the future. This should hopefully save us from
needing to do lots of maintenance on configure for the exotic
distributions out there who put libs in the most exotic places.