When using multiple catalogs always the first is chosen.
With the introduction of the autoselection of the right catalog when
none is selected we also auto select a catalog in the pool resource.
That is of course wrong as that means every pool always has an
explicit override of the catalog to use to the first catalog which
makes the override in the client useless.

Fixes #229: When using multiple catalogs always the first is chosen.