Fix some bugs in the NDMP implementation.
Up until now we were not fully RFC compliant. You need to restore the full NDMP
environment returned during a NDMP backup when performing a NDMP restore. We more
or less re-created the environment when doing a restore but were missing some important
elements. We now save the full environment in the database and use that to restore
the whole environment when doing a restore.