Remove hardcoded job timeout of 6 days.
The current watchdog setup doesn't make any sense with the state the
software is in 2014. So changed the watchdog timeout to be set to zero
by default which means don't run a jcr watchdog thread in either the
filed, stored or dird at all. You can still set an absolute timeout
on the job length but in normal situations it makes little sense.